Fix int/size_t, Makefile for cpp

+++ src/cpp_files/obj_structs.hpp
@@ -4,18 +4,18 @@
 namespace s21 {
 typedef struct Matrix {
   double** matrix = nullptr;
-  int rows = 0;
-  int cols = 0;
+  size_t rows = 0;
+  size_t cols = 0;
 } matrix_t;
 
 typedef struct facets {
-  int* vertexes = nullptr;
-  int number_of_vertexes_in_facets = 0;
+  size_t* vertexes = nullptr;
+  size_t number_of_vertexes_in_facets = 0;
 } polygon_t;
 
 typedef struct data {
-  int count_of_vertexes = 0;
-  int count_of_facets = 0;
+  size_t count_of_vertexes = 0;
+  size_t count_of_facets = 0;
   matrix_t matrix_3d;
   polygon_t* polygons = nullptr;
 } data;
